link up homepage with routes
